+++ checker/src/main/java/org/checkerframework/checker/nullness/KeyForAnalysis.java
@@ -1,11 +1,10 @@
 package org.checkerframework.checker.nullness;
 
-import java.util.Set;
-import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.type.TypeMirror;
 import org.checkerframework.common.basetype.BaseTypeChecker;
 import org.checkerframework.framework.flow.CFAbstractAnalysis;
 import org.checkerframework.framework.flow.CFAbstractValue;
+import org.checkerframework.javacutil.AnnotationMirrorSet;
 
 /** Boilerplate code to glue together all the parts the KeyFor dataflow classes. */
 public class KeyForAnalysis extends CFAbstractAnalysis<KeyForValue, KeyForStore, KeyForTransfer> {
@@ -32,7 +31,7 @@ public class KeyForAnalysis extends CFAbstractAnalysis<KeyForValue, KeyForStore,
 
   @Override
   public KeyForValue createAbstractValue(
-      Set<AnnotationMirror> annotations, TypeMirror underlyingType) {
+      AnnotationMirrorSet annotations, TypeMirror underlyingType) {
 
     if (!CFAbstractValue.validateSet(annotations, underlyingType, qualifierHierarchy)) {
       return null;