+++ framework/src/main/java/org/checkerframework/framework/stub/AnnotationFileElementTypes.java
@@ -21,14 +21,12 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.StringJoiner;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import javax.annotation.processing.ProcessingEnvironment;
-import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ElementKind;
 import javax.lang.model.element.ExecutableElement;
@@ -47,6 +45,7 @@ import org.checkerframework.framework.stub.AnnotationFileUtil.AnnotationFileType
 import org.checkerframework.framework.type.AnnotatedTypeFactory;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
 import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedExecutableType;
+import org.checkerframework.javacutil.AnnotationMirrorSet;
 import org.checkerframework.javacutil.BugInCF;
 import org.checkerframework.javacutil.ElementUtils;
 import org.checkerframework.javacutil.Pair;
@@ -398,7 +397,7 @@ public class AnnotationFileElementTypes {
    * @deprecated use {@link #getDeclAnnotations}
    */
   @Deprecated // 2021-06-26
-  public Set<AnnotationMirror> getDeclAnnotation(Element elt) {
+  public AnnotationMirrorSet getDeclAnnotation(Element elt) {
     return getDeclAnnotations(elt);
   }
 
@@ -412,9 +411,9 @@ public class AnnotationFileElementTypes {
    *     the annotation file and in the element. {@code null} is returned if {@code element} does
    *     not appear in an annotation file.
    */
-  public Set<AnnotationMirror> getDeclAnnotations(Element elt) {
+  public AnnotationMirrorSet getDeclAnnotations(Element elt) {
     if (isParsing()) {
-      return Collections.emptySet();
+      return AnnotationMirrorSet.emptySet();
     }
 
     parseEnclosingJdkClass(elt);
@@ -460,7 +459,7 @@ public class AnnotationFileElementTypes {
         }
       }
     }
-    return Collections.emptySet();
+    return AnnotationMirrorSet.emptySet();
   }
 
   /**