+++ framework/src/main/java/org/checkerframework/framework/type/treeannotator/PropagationTreeAnnotator.java
@@ -21,7 +21,7 @@ import org.checkerframework.framework.type.AnnotatedTypeMirror;
 import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedArrayType;
 import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedExecutableType;
 import org.checkerframework.framework.type.QualifierHierarchy;
-import org.checkerframework.javacutil.AnnotationUtils;
+import org.checkerframework.javacutil.AnnotationMirrorSet;
 import org.checkerframework.javacutil.CollectionUtils;
 import org.checkerframework.javacutil.Pair;
 import org.checkerframework.javacutil.TreePathUtil;
@@ -85,7 +85,7 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
       for (ExpressionTree init : tree.getInitializers()) {
         AnnotatedTypeMirror initType = atypeFactory.getAnnotatedType(init);
         // initType might be a typeVariable, so use effectiveAnnotations.
-        Set<AnnotationMirror> annos = initType.getEffectiveAnnotations();
+        AnnotationMirrorSet annos = initType.getEffectiveAnnotations();
 
         prev = (prev == null) ? annos : qualHierarchy.leastUpperBounds(prev, annos);
       }
@@ -182,14 +182,14 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
   }
 
   @Override
-  public Void visitCompoundAssignment(CompoundAssignmentTree node, AnnotatedTypeMirror type) {
+  public Void visitCompoundAssignment(CompoundAssignmentTree tree, AnnotatedTypeMirror type) {
     if (hasPrimaryAnnotationInAllHierarchies(type)) {
       // If the type already has a primary annotation in all hierarchies, then the
       // propagated annotations won't be applied.  So don't compute them.
       return null;
     }
-    AnnotatedTypeMirror rhs = atypeFactory.getAnnotatedType(node.getExpression());
-    AnnotatedTypeMirror lhs = atypeFactory.getAnnotatedType(node.getVariable());
+    AnnotatedTypeMirror rhs = atypeFactory.getAnnotatedType(tree.getExpression());
+    AnnotatedTypeMirror lhs = atypeFactory.getAnnotatedType(tree.getVariable());
     Set<? extends AnnotationMirror> lubs =
         qualHierarchy.leastUpperBounds(
             rhs.getEffectiveAnnotations(), lhs.getEffectiveAnnotations());
@@ -198,7 +198,7 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
   }
 
   @Override
-  public Void visitBinary(BinaryTree node, AnnotatedTypeMirror type) {
+  public Void visitBinary(BinaryTree tree, AnnotatedTypeMirror type) {
     if (hasPrimaryAnnotationInAllHierarchies(type)) {
       // If the type already has a primary annotation in all hierarchies, then the
       // propagated annotations won't be applied.  So don't compute them.
@@ -207,7 +207,7 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
       return null;
     }
 
-    Pair<AnnotatedTypeMirror, AnnotatedTypeMirror> argTypes = atypeFactory.binaryTreeArgTypes(node);
+    Pair<AnnotatedTypeMirror, AnnotatedTypeMirror> argTypes = atypeFactory.binaryTreeArgTypes(tree);
     Set<? extends AnnotationMirror> lubs =
         qualHierarchy.leastUpperBounds(
             argTypes.first.getEffectiveAnnotations(), argTypes.second.getEffectiveAnnotations());
@@ -217,14 +217,14 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
   }
 
   @Override
-  public Void visitUnary(UnaryTree node, AnnotatedTypeMirror type) {
+  public Void visitUnary(UnaryTree tree, AnnotatedTypeMirror type) {
     if (hasPrimaryAnnotationInAllHierarchies(type)) {
       // If the type already has a primary annotation in all hierarchies, then the
       // propagated annotations won't be applied.  So don't compute them.
       return null;
     }
 
-    AnnotatedTypeMirror exp = atypeFactory.getAnnotatedType(node.getExpression());
+    AnnotatedTypeMirror exp = atypeFactory.getAnnotatedType(tree.getExpression());
     type.addMissingAnnotations(exp.getAnnotations());
     return null;
   }
@@ -232,25 +232,25 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
   /*
    * TODO: would this make sense in general?
   @Override
-  public Void visitConditionalExpression(ConditionalExpressionTree node, AnnotatedTypeMirror type) {
+  public Void visitConditionalExpression(ConditionalExpressionTree tree, AnnotatedTypeMirror type) {
       if (!type.isAnnotated()) {
-          AnnotatedTypeMirror a = typeFactory.getAnnotatedType(node.getTrueExpression());
-          AnnotatedTypeMirror b = typeFactory.getAnnotatedType(node.getFalseExpression());
-          Set<AnnotationMirror> lubs = qualHierarchy.leastUpperBounds(a.getEffectiveAnnotations(), b.getEffectiveAnnotations());
+          AnnotatedTypeMirror a = typeFactory.getAnnotatedType(tree.getTrueExpression());
+          AnnotatedTypeMirror b = typeFactory.getAnnotatedType(tree.getFalseExpression());
+          AnnotationMirrorSet lubs = qualHierarchy.leastUpperBounds(a.getEffectiveAnnotations(), b.getEffectiveAnnotations());
           type.replaceAnnotations(lubs);
       }
-      return super.visitConditionalExpression(node, type);
+      return super.visitConditionalExpression(tree, type);
   }*/
 
   @Override
-  public Void visitTypeCast(TypeCastTree node, AnnotatedTypeMirror type) {
+  public Void visitTypeCast(TypeCastTree tree, AnnotatedTypeMirror type) {
     if (hasPrimaryAnnotationInAllHierarchies(type)) {
       // If the type is already has a primary annotation in all hierarchies, then the
       // propagated annotations won't be applied.  So don't compute them.
       return null;
     }
 
-    AnnotatedTypeMirror exprType = atypeFactory.getAnnotatedType(node.getExpression());
+    AnnotatedTypeMirror exprType = atypeFactory.getAnnotatedType(tree.getExpression());
     if (type.getKind() == TypeKind.TYPEVAR) {
       if (exprType.getKind() == TypeKind.TYPEVAR) {
         // If both types are type variables, take the direct annotations.
@@ -259,7 +259,7 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
       // else do nothing.
     } else {
       // Use effective annotations from the expression, to get upper bound of type variables.
-      Set<AnnotationMirror> expressionAnnos = exprType.getEffectiveAnnotations();
+      AnnotationMirrorSet expressionAnnos = exprType.getEffectiveAnnotations();
 
       TypeKind castKind = type.getPrimitiveKind();
       if (castKind != null) {
@@ -308,9 +308,9 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
    * @param annos annotations to add to type
    */
   private void addAnnoOrBound(AnnotatedTypeMirror type, Set<? extends AnnotationMirror> annos) {
-    Set<AnnotationMirror> boundAnnos =
+    AnnotationMirrorSet boundAnnos =
         atypeFactory.getQualifierUpperBounds().getBoundQualifiers(type.getUnderlyingType());
-    Set<AnnotationMirror> annosToAdd = AnnotationUtils.createAnnotationSet();
+    AnnotationMirrorSet annosToAdd = new AnnotationMirrorSet();
     for (AnnotationMirror boundAnno : boundAnnos) {
       AnnotationMirror anno = qualHierarchy.findAnnotationInSameHierarchy(annos, boundAnno);
       if (anno != null && !qualHierarchy.isSubtype(anno, boundAnno)) {